rename directory and change database
