+++ package.json
@@ -32,6 +32,7 @@
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
     "mysql": "^2.18.1",
+    "pg": "^8.12.0",
     "reflect-metadata": "^0.2.2",
     "tsyringe": "^4.8.0",
     "typescript": "^5.5.3"
