+++ src/domain/publications/interface/service/IServiceUseCase.ts
