+++ src/domain/publications/repositories/ServiceRepository.ts
@@ -1,4 +1,4 @@
-import AppDataSource from '@infrastructure/database/mysql/AppDataSource'
+import AppDataSource from '@infrastructure/database/psql/AppDataSource'
 import { injectable } from 'tsyringe'
 import { Repository } from 'typeorm'
 import { Service } from '../entities/services.entity'
