+++ src/ormconfig.ts
@@ -5,7 +5,7 @@ import dotenv from 'dotenv'
 dotenv.config()
 
 export default {
-  type: 'mysql',
+  type: 'postgres',
   host: process.env.DB_CONFIG_HOST,
   database: process.env.DB_CONFIG_DATABASE,
   port: process.env.DB_CONFIG_PORT,
@@ -15,9 +15,9 @@ export default {
   dropSchema: false,
   entities: [__dirname + '/**/*.entity{.ts,.js}'],
   migrations: [
-    __dirname + '/infrastructure/database/mysql/migrations/*{.ts,.js}',
+    __dirname + '/infrastructure/database/psql/migrations/*{.ts,.js}',
   ],
   cli: {
-    migrationsDir: __dirname + '/infrastructure/database/mysql/migrations',
+    migrationsDir: __dirname + '/infrastructure/database/psql/migrations',
   },
 } as DataSourceOptions
