+++ src/presentation/http/controllers/publications/services/ListServicesController.ts
@@ -3,20 +3,20 @@ import IController from '@shared/interfaces/IController'
 import { Request, Response } from 'express'
 import { inject, injectable } from 'tsyringe'
 import BaseController from '../../BaseController'
-import IServiceService from '@domain/publications/interface/service/IServiceServices'
+import IServiceUseCase from '@domain/publications/interface/service/IServiceUseCase'
 
 @injectable()
 export class ListServicesController extends BaseController implements IController {
   constructor(
-    @inject(tokens.ServiceServices)
-    private servicesService: IServiceService
+    @inject(tokens.ServiceUseCase)
+    private serviceUseCase: IServiceUseCase
   ) {
     super()
   }
 
   async handle(_req: Request, res: Response): Promise<Response> {
     try {
-      const services = await this.servicesService.list()
+      const services = await this.serviceUseCase.list()
       return this.success(res, 'Lista de Serviços', services)
     } catch (err: any) {
       return this.error(res, err.message)
